Changed Map to update current user location
Code was more or less correct from before. Issue was with trying to get
the user location too quickly after calling connect()
The fix was basically to move the code into the correct methods so they
ran in the right order and without running each over (calling for last
location before there is a location). We now create the locationClient
in onCreate(), connect the client in onStart(), and update the user
location and draw the map in onConnected() (makes intuitive sense
although it wasn't immediately apparent). I believe you should be able
to see the differences in the code highlighted in red for parts that
have been deleted and green where they've been added.