GamePaused/Map
Found bug where markers were being drawn in multiple locations for the
user after they left map screen and returned. This happened because
onConnected is called every time the user returns to the map screen (ie
after leaving the pause or store activities) and our draw method is in
onConnected. I added a map.clear() call in onConnected to clear the
marker before being redrawn with the users new location.

GamePaused was only changed to make the resume button perform the same
action as the back button (probably not ideal but better than going to
store for now)